requirements.txt
fastapi
uvicorn
pydantic
tensorflow
pandas
numpy
streamlit
requests
Pillow
openai